+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog is following the recommended format by [keepachangelog](https://k
 ### Added
 
 - Add support for cloud rules by [@animesh-tarodia-sp](https://github.com/animesh-tarodia-sp) (cf. [#163](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/pull/163))
+- Export a cloud rule or a `.java`/`.bsh` file as XML
 
 ### Changed